dashboard: consolidate Rewards into Payouts (remove Rewards page)
Merge the Rewards page into Payouts. The 'Where Rewards Come From'
revenue-streams explainer and the 'Node Balance Accounts' per-node
balance table now live on the Payouts page. The 'Share Contribution'
section is dropped as redundant with the Capabilities page.

Remove the Rewards route and its sidebar nav item (the now-empty
'operate' group and unused Coins icon import go with it), and repoint
the Ctrl+R keyboard shortcut to /payouts.
